requirements.txt
alembic==1.16.4
greenlet==3.2.3
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pytesseract==0.3.13
python-doctr[torch]==1.0.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
